oops, botched merge
git-svn-id: svn+ssh://corpsvnproxy/yahoo/BrowserPlus/public_platform/trunk@349 3fcf768e-b076-0410-b9b8-e8c34e2d470b

+++ src/libs/bputils/src/Http/Client/HttpTransaction_Windows.cpp
@@ -905,7 +905,7 @@ Transaction::Impl::sendRequestWithBody()
             setError("unable to allocate buffer");
             return GetLastError();
         }
-        m_sendTotalBytes = bp::file::size(path);
+        m_sendTotalBytes = (size_t) bp::file::size(path);
 	} else {
         m_ePostSource = eFromBuffer;
         m_sendTotalBytes = m_pRequest->body.size();
